Move into Carla, also the CarlaActionWrapper
